cold-person/pyproject.toml
[project]
name        = "cold-person"
description = "A person who might want a sweater"
version     = "0.1.0"

dependencies = [
  "typing-extensions ~= 4.0 ; python_version < '3.10'",
  "aiohttp[speedups] ~= 3.0",
]
[project.optional-dependencies]
test = [
  "pytest            ~= 7.0",
  "pytest-asyncio    ~= 0.0",
  "pytest-aiohttp    ~= 1.0",
]
dev = [
  "pact-python[types]",
  "pact-python[test]",
  "black          ==23.12.1",
  "flake8         ~= 7.0",
  "isort          ~= 5.0",
]

[project.scripts]
coldperson-serve = "coldperson.cli.serve:main"

[tool.pytest.ini_options]
minversion = "6.0"
asyncio_mode = "strict"
testpaths = [
  "tests",
]
